Added home mode tracker, that sets halo backlight on or off.
Added second mass and handling it's commands and events.

+++ src/internal/transformer/strategies_test.go
@@ -56,7 +56,7 @@ func TestFimpStrategies(t *testing.T) {
 			val:      10,
 			device: config.Device{
 				Config: map[string]interface{}{
-					"minValue": 10,
+					"minValue": float64(10),
 				},
 			},
 			want: presence(t, false),
